chatbot_api/pyproject.toml
[project]
name = "chatbot_api"
version = "0.1"
dependencies = [
    "fastapi==0.115.6",
    "langchain==0.3.13",
   " langchain_community==0.3.13",
   " langchain_core==0.3.28",
    "langchain_groq==0.2.2",
    "langchain_huggingface==0.1.2",
    "langchain_neo4j==0.2.0",
    "langgraph==0.2.60",
    "numpy==1.26.4",
    "pydantic==2.10.4",
    "python-dotenv==1.0.1",
    "uvicorn==0.15.0"
]
[project.optional-dependencies]
dev = ["black", "flake8"]